fix: remove stale messages, rename the cli_common to tasks
